refactor: streamline scanning logic and remove redundant code
Consolidate and simplify scan handling across client and server by
removing duplicated logic, clarifying responsibilities, and improving
data flow. This reduces maintenance overhead, fixes subtle edge-cases in
scan task state updates, and makes the delta reporting workflow easier
to follow.
