+++ server/app/models/scan_task.py
@@ -102,161 +102,6 @@ class ScanTask(db.Model):
 
         return all_completed
 
-    def get_task_group_tasks(self):
-        """
-        Get all tasks in this task group.
-
-        Returns:
-            list: List of ScanTask objects in the same group
-        """
-        return (
-            ScanTask.query.filter_by(task_group_id=self.task_group_id)
-            .order_by(ScanTask.created_at)
-            .all()
-        )
-
-    def get_task_group_results(self):
-        """
-        Get all scan results from tasks in this group.
-
-        Returns:
-            list: List of ScanResult objects from this task group
-        """
-        from app.models.scan_result import ScanResult
-
-        group_tasks = self.get_task_group_tasks()
-
-        # Get all scan results that have a completed status
-        results = []
-        for task in group_tasks:
-            if task.scan_result_id:
-                result = ScanResult.query.get(task.scan_result_id)
-                if result and result.status == "completed":
-                    results.append(result)
-
-        return results
-
-    def get_previous_task_group_id(self):
-        """
-        Find the task_group_id of the previous completed task group for this scan.
-
-        Returns:
-            str: Previous task_group_id or None if this is the first
-        """
-        # Find all task groups for this scan that completed before this one
-        previous_groups = (
-            db.session.query(
-                ScanTask.task_group_id,
-                db.func.max(ScanTask.completed_at).label("max_completed"),
-            )
-            .filter(
-                ScanTask.scan_id == self.scan_id,
-                ScanTask.task_group_id != self.task_group_id,
-                ScanTask.status == "completed",
-                ScanTask.completed_at < self.completed_at,
-            )
-            .group_by(ScanTask.task_group_id)
-            .order_by(db.desc("max_completed"))
-            .first()
-        )
-
-        if previous_groups:
-            return previous_groups.task_group_id
-
-        return None
-
-    @staticmethod
-    def aggregate_scan_results(scan_results):
-        """
-        Aggregate multiple scan results into a single results_data structure.
-        Combines hosts and ports from all results.
-
-        Args:
-            scan_results: List of ScanResult objects
-
-        Returns:
-            dict: Aggregated results_data in the format expected by compare_with()
-        """
-        aggregated = {"hosts": {}}
-
-        for result in scan_results:
-            if not result.results_data:
-                continue
-
-            result_hosts = result.results_data.get("hosts", {})
-
-            for host_ip, host_data in result_hosts.items():
-                if host_ip not in aggregated["hosts"]:
-                    # First time seeing this host
-                    aggregated["hosts"][host_ip] = {"ports": []}
-
-                # Merge ports (avoid duplicates)
-                existing_ports = {
-                    p["port"]: p for p in aggregated["hosts"][host_ip]["ports"]
-                }
-
-                for port_info in host_data.get("ports", []):
-                    port_num = port_info.get("port")
-                    if port_num:
-                        # Keep the most recent/complete information
-                        if port_num not in existing_ports or len(
-                            str(port_info.get("service", ""))
-                        ) > len(str(existing_ports[port_num].get("service", ""))):
-                            existing_ports[port_num] = port_info
-
-                # Convert back to list
-                aggregated["hosts"][host_ip]["ports"] = list(existing_ports.values())
-
-        return aggregated
-
-    @staticmethod
-    def create_task_group(
-        scan_id, client_ranges, ports="1-1000", scan_type="tcp", priority=5
-    ):
-        """
-        Create a group of related scan tasks for multiple clients.
-        All tasks in the group share the same task_group_id.
-
-        Args:
-            scan_id: The scan ID these tasks belong to
-            client_ranges: List of dicts with 'client_id' and 'targets'
-                          Example: [
-                              {'client_id': 'client1', 'targets': ['192.168.1.0/25']},
-                              {'client_id': 'client2', 'targets': ['192.168.1.128/25']}
-                          ]
-            ports: Port specification
-            scan_type: Type of scan
-            priority: Priority level
-
-        Returns:
-            tuple: (task_group_id, list of created ScanTask objects)
-        """
-        # Generate a unique task_group_id for this execution
-        task_group_id = str(uuid.uuid4())
-
-        tasks = []
-
-        for client_range in client_ranges:
-            task = ScanTask(
-                task_id=str(uuid.uuid4()),
-                task_group_id=task_group_id,
-                client_id=client_range.get("client_id"),
-                scan_id=scan_id,
-                targets=json.dumps(client_range.get("targets", [])),
-                ports=ports,
-                scan_type=scan_type,
-                priority=priority,
-                status="pending",
-            )
-            tasks.append(task)
-            db.session.add(task)
-
-        db.session.commit()
-
-        print(f"✓ Created task group {task_group_id} with {len(tasks)} tasks")
-
-        return task_group_id, tasks
-
     def to_dict(self):
         """Convert scan task to dictionary"""
         return {
