+++ server/app/routes/main.py
@@ -3,7 +3,6 @@ Main routes for the Port Scanner Delta Reporter
 """
 
 from flask import Blueprint, render_template
-from app.forms import LoginForm
 
 
 bp = Blueprint("main", __name__)
@@ -13,17 +12,3 @@ bp = Blueprint("main", __name__)
 def index():
     """Home page"""
     return render_template("index.html")
-
-
-@bp.route("/about")
-def about():
-    """About page"""
-    return render_template("about.html")
-
-
-''' Example additional route
-@bp.route("/hello")
-def hello():
-    """About page"""
-    return render_template("index.html")
-'''
