+++ server/app/routes/scan.py
@@ -352,9 +352,6 @@ def delete_scan(scan_id):
         return jsonify({"status": "error", "message": str(e)}), 500
 
 
-# ============== SCAN EXECUTION ==============
-
-
 @bp.route("/scans/<int:scan_id>/execute", methods=["POST"])
 @login_required
 def execute_scan(scan_id):
@@ -427,112 +424,4 @@ def toggle_scan(scan_id):
     except Exception as e:
         db.session.rollback()
         current_app.logger.error(f"Failed to toggle scan: {str(e)}")
-        return jsonify({"status": "error", "message": str(e)}), 500
-
-
-@bp.route("/scans/<int:scan_id>/schedule", methods=["POST"])
-@login_required
-def update_scan_schedule(scan_id):
-    """Update scan scheduling"""
-    try:
-        # Get scan with access control
-        if current_user.is_admin:
-            scan = Scan.query.get(scan_id)
-        else:
-            scan = Scan.query.filter_by(id=scan_id, user_id=current_user.id).first()
-
-        if not scan:
-            return jsonify({"status": "error", "message": "Scan not found"}), 404
-
-        data = request.get_json()
-
-        # Update scheduling parameters
-        if "is_scheduled" in data:
-            scan.is_scheduled = data["is_scheduled"]
-        if "interval_minutes" in data:
-            scan.interval_minutes = data["interval_minutes"]
-
-        scan.updated_at = datetime.utcnow()
-
-        # Update scheduler
-        if scan.is_scheduled and scan.is_active:
-            scan.next_run = datetime.utcnow() + timedelta(minutes=scan.interval_minutes)
-            scheduler_service.schedule_scan(scan)
-        else:
-            scan.next_run = None
-            _unschedule_scan(scan_id)
-
-        db.session.commit()
-
-        return (
-            jsonify(
-                {
-                    "status": "success",
-                    "message": "Schedule updated successfully",
-                    "scan": scan.to_dict(),
-                }
-            ),
-            200,
-        )
-
-    except Exception as e:
-        db.session.rollback()
-        current_app.logger.error(f"Failed to update schedule: {str(e)}")
-        return jsonify({"status": "error", "message": str(e)}), 500
-
-
-# ============== SCAN RESULTS ==============
-
-
-@bp.route("/scans/<int:scan_id>/results", methods=["GET"])
-@login_required
-def get_scan_results(scan_id):
-    """Get results for a specific scan"""
-    try:
-        # Verify access
-        if current_user.is_admin:
-            scan = Scan.query.get(scan_id)
-        else:
-            scan = Scan.query.filter_by(id=scan_id, user_id=current_user.id).first()
-
-        if not scan:
-            return jsonify({"status": "error", "message": "Scan not found"}), 404
-
-        # Query parameters
-        limit = request.args.get("limit", 20, type=int)
-        offset = request.args.get("offset", 0, type=int)
-        status = request.args.get("status")  # completed, failed, running
-
-        # Build query
-        query = ScanResult.query.filter(
-            ScanResult.results_data.contains({"scan_config_id": scan_id})
-        )
-
-        if status:
-            query = query.filter_by(status=status)
-
-        # Order by creation date (newest first)
-        query = query.order_by(ScanResult.created_at.desc())
-
-        # Get total count
-        total = query.count()
-
-        # Apply pagination
-        results = query.limit(limit).offset(offset).all()
-
-        return (
-            jsonify(
-                {
-                    "status": "success",
-                    "total": total,
-                    "limit": limit,
-                    "offset": offset,
-                    "results": [result.to_dict() for result in results],
-                }
-            ),
-            200,
-        )
-
-    except Exception as e:
-        current_app.logger.error(f"Failed to get scan results: {str(e)}")
         return jsonify({"status": "error", "message": str(e)}), 500
