+++ server/app/templates/base.html
@@ -7,6 +7,10 @@
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
       rel="stylesheet" />
+    <link
+      href=" https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css "
+      rel="stylesheet" />
+
     <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}" />
     {% block extra_css %}{% endblock %}
   </head>
