+++ server/app/templates/dashboard/view_report.html
@@ -1,0 +1,167 @@
+{% extends "base.html" %}
+
+{% block title %}Delta Reports - {{ scan.name }}{% endblock %}
+
+{% block extra_css %}
+<link rel="stylesheet" href="{{ url_for('static', filename='css/delta_reports.css') }}">
+{% endblock %}
+
+{% block content %}
+<div class="container-fluid mt-4">
+    <!-- Header Section -->
+    <div class="row mb-4">
+        <div class="col-md-12">
+            <nav aria-label="breadcrumb">
+                <ol class="breadcrumb">
+                    <li class="breadcrumb-item"><a href="{{ url_for('scan.list_scans') }}">Scans</a></li>
+                    <li class="breadcrumb-item"><a href="{{ url_for('scan.get_scan', scan_id=scan.id) }}">{{ scan.name }}</a></li>
+                    <li class="breadcrumb-item active" aria-current="page">Delta Reports</li>
+                </ol>
+            </nav>
+
+            <div class="d-flex justify-content-between align-items-center">
+                <div>
+                    <h2 class="mb-1">
+                        <i class="fas fa-chart-line"></i> Change History
+                    </h2>
+                    <p class="text-muted">{{ scan.name }}</p>
+                </div>
+                <div>
+                    <button class="btn btn-outline-primary" id="refreshBtn">
+                        <i class="fas fa-sync-alt"></i> Refresh
+                    </button>
+                    <button class="btn btn-outline-secondary" id="exportAllBtn">
+                        <i class="fas fa-download"></i> Export All
+                    </button>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <!-- Filters Section -->
+    <div class="row mb-4">
+        <div class="col-md-12">
+            <div class="card">
+                <div class="card-body">
+                    <div class="row align-items-center">
+                        <div class="col-md-6">
+                            <div class="form-check form-switch">
+                                <input class="form-check-input" type="checkbox" id="onlyChangesFilter">
+                                <label class="form-check-label" for="onlyChangesFilter">
+                                    Only show reports with changes
+                                </label>
+                            </div>
+                        </div>
+                        <div class="col-md-3">
+                            <select class="form-select" id="perPageSelect">
+                                <option value="10">10 per page</option>
+                                <option value="25">25 per page</option>
+                                <option value="50">50 per page</option>
+                                <option value="100">100 per page</option>
+                            </select>
+                        </div>
+                        <div class="col-md-3">
+                            <div id="reportCount" class="text-muted text-end">
+                                Loading...
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <!-- Loading Spinner -->
+    <div id="loadingSpinner" class="text-center py-5" style="display: none;">
+        <div class="spinner-border text-primary" role="status">
+            <span class="visually-hidden">Loading...</span>
+        </div>
+        <p class="mt-3 text-muted">Loading delta reports...</p>
+    </div>
+
+    <!-- Reports List -->
+    <div id="reportsList" class="row">
+        <!-- Reports will be dynamically inserted here -->
+    </div>
+
+    <!-- Pagination -->
+    <div id="paginationContainer" class="row mt-4" style="display: none;">
+        <div class="col-md-12">
+            <nav aria-label="Delta reports pagination">
+                <ul class="pagination justify-content-center" id="pagination">
+                    <!-- Pagination will be dynamically inserted here -->
+                </ul>
+            </nav>
+        </div>
+    </div>
+
+    <!-- Empty State -->
+    <div id="emptyState" class="text-center py-5" style="display: none;">
+        <i class="fas fa-inbox fa-4x text-muted mb-3"></i>
+        <h4 class="text-muted">No Delta Reports Found</h4>
+        <p class="text-muted" id="emptyStateMessage">
+            Delta reports will appear here after your second scan completes.
+        </p>
+    </div>
+</div>
+
+<!-- Report Detail Modal -->
+<div class="modal fade" id="reportDetailModal" tabindex="-1" aria-labelledby="reportDetailModalLabel" aria-hidden="true">
+    <div class="modal-dialog modal-xl modal-dialog-scrollable">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="reportDetailModalLabel">Delta Report Details</h5>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div class="modal-body" id="modalBodyContent">
+                <div class="text-center py-5">
+                    <div class="spinner-border text-primary" role="status">
+                        <span class="visually-hidden">Loading...</span>
+                    </div>
+                </div>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                <button type="button" class="btn btn-primary" id="exportModalReportBtn">
+                    <i class="fas fa-download"></i> Export CSV
+                </button>
+            </div>
+        </div>
+    </div>
+</div>
+
+<!-- Delete Confirmation Modal -->
+<div class="modal fade" id="deleteConfirmModal" tabindex="-1" aria-labelledby="deleteConfirmModalLabel" aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header bg-danger text-white">
+                <h5 class="modal-title" id="deleteConfirmModalLabel">Confirm Delete</h5>
+                <button type="button" class="btn-close btn-close-white" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div class="modal-body">
+                <p>Are you sure you want to delete this delta report?</p>
+                <p class="text-muted small mb-0">This action cannot be undone.</p>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
+                <button type="button" class="btn btn-danger" id="confirmDeleteBtn">
+                    <i class="fas fa-trash"></i> Delete
+                </button>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
+
+{% block extra_js %}
+<script src="{{ url_for('static', filename='js/delta_reports.js') }}"></script>
+<script>
+    // Initialize with scan ID from template
+    const SCAN_ID = {{ scan.id }};
+
+    // Initialize the delta reports page
+    document.addEventListener('DOMContentLoaded', function() {
+        DeltaReports.init(SCAN_ID);
+    });
+</script>
+{% endblock %}
